requirements.txt
fastapi
uvicorn
streamlit
sqlalchemy
psycopg2-binary
google-generativeai
python-dotenv